requirements.txt
fastapi
uvicorn
torch
numpy
opencv-python
PyYAML
requests